Use new models from migrations over legacy models
