+++ src/models/index.js
@@ -1,34 +1,47 @@
-import Sequelize from 'sequelize';
+/* eslint-disable import/no-dynamic-require */
 
-let sequelize; // eslint-disable-line import/no-mutable-exports
+const fs = require('fs');
+const path = require('path');
+const Sequelize = require('sequelize');
 
-if (process.env.NODE_ENV === 'production') {
-  sequelize = new Sequelize(process.env.DATABASE_URL, {
-    dialect: 'postgres',
-    dialectOptions: {
-      ssl: true,
-    },
-  });
+const basename = path.basename(__filename);
+const env = process.env.NODE_ENV || 'development';
+const config = require(`${__dirname}/../../config/config.js`)[env];
+const db = {};
+
+let sequelize;
+if (config.use_env_variable) {
+  sequelize = new Sequelize(process.env[config.use_env_variable], config);
 } else {
   sequelize = new Sequelize(
-    process.env.DATABASE,
-    process.env.DATABASE_USER,
-    process.env.DATABASE_PASSWORD,
-    {
-      dialect: 'postgres',
-    },
+    config.database,
+    config.username,
+    config.password,
+    config,
   );
 }
 
-const models = {
-  Note: sequelize.import('./note'),
-  User: sequelize.import('./user'),
-};
+fs.readdirSync(__dirname)
+  .filter(
+    (file) =>
+      file.indexOf('.') !== 0 && file !== basename && file.slice(-3) === '.js',
+  )
+  .forEach((file) => {
+    // eslint-disable-next-line global-require
+    const model = require(path.join(__dirname, file))(
+      sequelize,
+      Sequelize.DataTypes,
+    );
+    db[model.name] = model;
+  });
 
-Object.keys(models).forEach((model) => {
-  models[model].associate(models);
+Object.keys(db).forEach((modelName) => {
+  if (db[modelName].associate) {
+    db[modelName].associate(db);
+  }
 });
 
-export { sequelize };
+db.sequelize = sequelize;
+db.Sequelize = Sequelize;
 
-export default models;
+module.exports = db;
